+++ src/adapter/graphql/tests/utils/base_gql_dataset_harness.rs
@@ -54,9 +54,9 @@ impl BaseGQLDatasetHarness {
             .add::<DidGeneratorDefault>()
             .add_value(tenancy_config)
             .add::<DatabaseTransactionRunner>()
-            .add_builder(odf::dataset::DatasetStorageUnitLocalFs::builder().with_root(datasets_dir))
-            .bind::<dyn odf::DatasetStorageUnit, odf::dataset::DatasetStorageUnitLocalFs>()
-            .bind::<dyn odf::DatasetStorageUnitWriter, odf::dataset::DatasetStorageUnitLocalFs>()
+            .add_builder(odf::dataset::DatasetStorageUnitLocalFs::builder(
+                datasets_dir,
+            ))
             .add::<DatasetLfsBuilderDatabaseBackedImpl>()
             .add::<CreateDatasetFromSnapshotUseCaseImpl>()
             .add::<CreateDatasetUseCaseImpl>()